fixed two more signed/unsigned comparisons

+++ src/io/remotesteerer/test/parallel_mpi_4/pipetest.h
@@ -32,7 +32,7 @@ public:
         TS_ASSERT_EQUALS(pipe.steeringRequests[1], "set flow 6.9 100 120 110");
 
         TS_ASSERT_EQUALS(pipe.retrieveSteeringRequests().size(), unsigned(2));
-        TS_ASSERT_EQUALS(pipe.steeringRequests.size(), 0);
+        TS_ASSERT_EQUALS(pipe.steeringRequests.size(), size_t(0));
     }
 
     void testSyncSteeringFeedback()
@@ -51,7 +51,7 @@ public:
         TS_ASSERT_EQUALS(pipe.steeringFeedback.size(),           expectedSize);
         TS_ASSERT_EQUALS(pipe.copySteeringFeedback().size(),     expectedSize);
         TS_ASSERT_EQUALS(pipe.retrieveSteeringFeedback().size(), expectedSize);
-        TS_ASSERT_EQUALS(pipe.steeringFeedback.size(), 0);
+        TS_ASSERT_EQUALS(pipe.steeringFeedback.size(), size_t(0));
     }
 
     class Runner
